Extract csv export functionality to service
