+++ src/app/payment-hub/request-to-pay/incoming-request-export/incoming-request-export.component.html
@@ -85,7 +85,7 @@
         <mat-label for="querytype" class="form-control">Filter By</mat-label>
         <select
           id="querytype"
-          name="cars"
+          name="status"
           class="form-control col-sm-4"
           ngModel
           matNativeControl
