+++ src/app/payment-hub/request-to-pay/incoming-request-to-pay/incoming-request-to-pay.component.html
@@ -232,7 +232,7 @@
           id="filterBy"
           required
           ngModel
-          name="transactionid"
+          name="val"
           placeholder="comma-separated-list"
         />
       </mat-form-field>
